Style task name.  Refs #9006.
 * Color background of task name span on left like task bar on right.
 * Strike through name of closed tasks.

Changes to jsgantt.js just aply the passed class to the task name (it
was previously only on the task bar in the Gantt).

Change to tracjsgantt.py pass closed class when appropriate.

Change to tracjsgantt.css create span (name) and dif (task) styles.

The left text alignment is needed because this change adds gtask class
to the task name.  I may come back and fix that but this works.