Refs #9770: Re-implemented the patch by ebencheung, with minor changes. The images are now stored in a cache directory, rather than in the session data. ebencheung had described the issue in which session data was limited by Apache, and the graphs failed to render. Caching files to disk seems like a better approach and this is similar to how GraphvizPlugin works. Some code review will be done in the future to compare this implementation with that of the GraphvizPlugin.
Minor changes from the original patch are:
 * The cache directory can be specified in `trac.ini`. The default is `$TRACENV/cache/plantuml`.
 * A `_read_img_from_file` method was added.
 * Minor coding style changes.
 
The biggest concern I have with this implementation is that the cache directory is never cleaned. I'm not sure of the best way to handle this, but as a first step I'll take a look at if/how the GraphvizPlugin handles this issue.