Add private table creation.  Refs #9648.
I'm not sure that the name of the table, schedule, is sufficiently
obscure that it won't step on users or other plugins.

We're going to store the schedule in a private table.  We can use
IEnvironmentSetupParticipant methods to check for the table and create
as needed the first time the plugin is activated.

The general principles are at http://trac.edgewall.org/wiki/TracDev/PluginDevelopment/ExtensionPoints/trac.env.IEnvironmentSetupParticipant

The Trac schema is at
http://trac.edgewall.org/browser/trunk/trac/db_default.py for
comparison.  I modeled `schedule` on `ticket_change` (`int64` for
datetimes) and `ticket_custom` (index by `ticket` to get values).

Some of the detail code came out of `api.py` in git://git.sixnetio.com/git/misc_tools/trac_plugin_source/trac-subtickets-plugin.git.