Use numerical precedence levels for start/finish sources.  Refs #9648.
Before a date was either explicit or not.  Values in the database
(actual start/finish dates, previously scheduled values, and explicit
user-supplied values) were all explicit so a test for "better" between
an old schedule table value and a new actual value were not resolved.
Now each source of a date has a different precedence so we have enough
information to resolve a conflict.

The way dates were calculated, we should not have seen conflicts in
new values, only in the explicit ones.  This change allows an actual
start to be higher precedence than a previously-scheduled finish.
However, there is still the potential conflict between items of the
same precedence.

 * Two actual values should never be wrong (we can't finish before we start)
 * Two scheduled values should never be wrong if we never save bad values
 * If two user-supplied values are wrong, it's data entry, not
algorithm.  We could validate this away but we don't currently even
allow explicit start dates.